availability-service/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pytest==8.2.0
pytest-mock==3.14.0